# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "visionatrix-client",
    "version": "0.8.0.dev0"
  },
  "paths": {
...
    "/api/flows-sub-flows": {
      "get": {
        "summary": "Flows From",
        "description": "Retrieves a list of flows designed to post-process the results from other flows, filtering by the type\nof input they handle, either 'image' or 'video'. This endpoint is particularly useful for chaining workflows\nwhere the output of one flow becomes the input to another. It modifies the main flow's structure by adopting\nsub-flow's display name and selectively merging input parameters from the sub-flows into the main flow's parameters\nbased on matching names.",
        "operationId": "flows_from_api_flows_sub_flows_get",
        "parameters": [
          {
            "name": "input_type",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "image",
                "video"
              ],
              "type": "string",
              "title": "Input Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flow"
                  },
                  "title": "Response Flows From Api Flows Sub Flows Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/flow": {
      "put": {
        "summary": "Flow Install",
        "description": "Endpoint to initiate the installation of a flow based on its name. This endpoint requires admin privileges\nto perform the installation. If another flow installation is already in progress, it prevents a new\ninstallation to avoid conflicts, returning a 409 Conflict HTTP status.\n\nThis endpoint schedules a background task for the installation process using the specified flow name. It\nchecks the availability of the flow in the list of available flows and starts the installation if the flow\nis found. It ensures that no two installations can run concurrently.",
        "operationId": "flow_install_api_flow_put",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Name of the flow you wish to install",
              "title": "Name"
            },
            "description": "Name of the flow you wish to install"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful start of installation"
          },
          "404": {
            "description": "Flow not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Can't find `flow_name` flow."
                }
              }
            }
          },
          "409": {
            "description": "Another flow installation is in progress",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Another flow installation is in progress."
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Flow Install From File",
        "description": "Endpoint to initiate the installation of a flow from an uploaded file. This endpoint requires admin privileges\nto perform the installation. If another flow installation is already in progress, it prevents a parallel flow\ninstallation to avoid conflicts, returning a 409 Conflict HTTP status.\n\nThis endpoint schedules a background task for the installation process using the uploaded flow file.",
        "operationId": "flow_install_from_file_api_flow_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_flow_install_from_file_api_flow_post"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful start of installation"
          },
          "409": {
            "description": "Another flow installation is in progress",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Another flow installation is in progress."
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Flow Delete",
        "description": "Endpoint to delete an installed flow by its name. Requires administrative privileges to execute.\nThis endpoint will succeed even if the flow does not exist.",
        "operationId": "flow_delete_api_flow_delete",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Name of the flow you wish to delete",
              "title": "Name"
            },
            "description": "Name of the flow you wish to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Flow deleted successfully"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/flow-progress-install": {
      "get": {
        "summary": "Flow Progress Install Get",
        "description": "Retrieves the current installation progress of all flows from an in-memory dictionary. This endpoint\nreturns a dictionary showing the installation status for each flow.\n\nRequires administrative privileges.",
        "operationId": "flow_progress_install_get_api_flow_progress_install_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlowProgressInstall"
                  },
                  "title": "Response Flow Progress Install Get Api Flow Progress Install Get"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Flow Progress Install Delete",
        "description": "Deletes the installation progress entry for a specified flow.\n\nRequires administrative privileges.",
        "operationId": "flow_progress_install_delete_api_flow_progress_install_delete",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Name of the flow progress you wish to delete",
              "title": "Name"
            },
            "description": "Name of the flow progress you wish to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Flow progress installation entry deleted successfully"
          },
          "404": {
            "description": "Flow progress installation entry not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Can't find `flow_name`."
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/engine-interrupt": {
      "post": {
        "summary": "Engine Interrupt",
        "description": "Interrupts the currently executing task. This is primarily an internal function and should be used\ncautiously. For standard task management, prefer using the `task_queue_clear` or `tasks_queue_clear`\nendpoints. Requires administrative privileges to execute.",
        "operationId": "engine_interrupt_api_engine_interrupt_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "204": {
            "description": "Engine interrupt initiated successfully"
          },
          "401": {
            "description": "Unauthorized - Admin privilege required",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Admin privilege required"
                }
              }
            }
          }
        }
      }
    },
    "/api/shutdown": {
      "post": {
        "summary": "Shutdown",
        "description": "Shuts down the current instance of Vix. This endpoint queues a task to terminate the server process\nafter a short delay, ensuring any final operations can complete. Access is restricted to administrators only.",
        "operationId": "shutdown_api_shutdown_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "204": {
            "description": "Server shutdown initiated successfully"
          },
          "401": {
            "description": "Unauthorized - Admin privilege required",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Admin privilege required"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIResourceModel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Unique name of the model."
          },
          "save_path": {
            "type": "string",
            "title": "Save Path",
            "description": "Subpath where the model is stored within the local system."
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "URL from which the model can be downloaded."
          },
          "homepage": {
            "type": "string",
            "title": "Homepage",
            "description": "Webpage with detailed information about the model.",
            "default": ""
          },
          "hash": {
            "type": "string",
            "title": "Hash",
            "description": "SHA256 hash of the model file for integrity verification."
          },
          "regexes": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Regexes",
            "description": "List of regex patterns that dynamically resolve model details based on workflow configurations.",
            "default": []
          },
          "gated": {
            "type": "boolean",
            "title": "Gated",
            "description": "Flag showing is the model closed to public access",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "save_path",
          "url",
          "hash"
        ],
        "title": "AIResourceModel",
        "description": "Represents an AI model resource within a Flow.\n\nThis model provides a structured way to handle AI\nmodels that are integral to workflows, ensuring that each model can be dynamically\nresolved, downloaded, and verified before use."
      },
      "Body_flow_install_from_file_api_flow_post": {
        "properties": {
          "flow_file": {
            "type": "string",
            "format": "binary",
            "title": "Flow File"
          }
        },
        "type": "object",
        "required": [
          "flow_file"
        ],
        "title": "Body_flow_install_from_file_api_flow_post"
      },
...
      "Flow": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The unique identifier of the flow."
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "The user-friendly name of the flow."
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "A brief explanation of the flow's purpose and functionality.",
            "default": ""
          },
          "author": {
            "type": "string",
            "title": "Author",
            "description": "The creator or maintainer of the flow."
          },
          "homepage": {
            "type": "string",
            "title": "Homepage",
            "description": "A URL to the flow's homepage or the author's website.",
            "default": ""
          },
          "license": {
            "type": "string",
            "title": "License",
            "description": "The type of license under which the flow is made available.",
            "default": ""
          },
          "documentation": {
            "type": "string",
            "title": "Documentation",
            "description": "A URL linking to detailed documentation for the flow.",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Tags describing this flow.",
            "default": []
          },
          "sub_flows": {
            "items": {
              "$ref": "#/components/schemas/SubFlow"
            },
            "type": "array",
            "title": "Sub Flows",
            "description": "A list of subflows derived from this flow, allowing customization or extension.",
            "default": []
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/AIResourceModel"
            },
            "type": "array",
            "title": "Models",
            "description": "A list of models used by the ComfyUI workflow.",
            "default": []
          },
          "input_params": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Input Params",
            "description": "Initial set of parameters required to launch the flow, potentially modifiable by subflows."
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Internal version of the flow in major.minor format.",
            "default": ""
          },
          "requires": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requires",
            "description": "Required external workflow dependencies.",
            "default": []
          },
          "private": {
            "type": "boolean",
            "title": "Private",
            "description": "Whether the workflow is missing from the `FLOWS_CATALOG_URL`",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "display_name",
          "author",
          "input_params"
        ],
        "title": "Flow",
        "description": "Flows serve as add-ons to ComfyUI workflows, determining the parameters to be displayed and populated.\nThey also allow for the modification of ComfyUI workflow behavior based on incoming parameters."
      },
      "FlowProgressInstall": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the flow being installed."
          },
          "flow": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Flow"
              },
              {
                "type": "null"
              }
            ],
            "description": "Parsed information about Flow in Visionatrix format."
          },
          "flow_comfy": {
            "type": "object",
            "title": "Flow Comfy",
            "description": "The ComfyUI workflow which are installed."
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "description": "The current progress of the installation, ranging from 0 to 100."
          },
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Details of any error encountered during the installation process.",
            "default": ""
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At",
            "description": "Timestamp when the installation process started."
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "Timestamp of the last update to the installation progress."
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At",
            "description": "Timestamp when the installation process completed."
          }
        },
        "type": "object",
        "required": [
          "name",
          "flow_comfy",
          "progress",
          "started_at"
        ],
        "title": "FlowProgressInstall",
        "description": "Represents the progress status of a flow installation process."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SubFlow": {
        "properties": {
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "The new display name when this subflow's parameters are used."
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "The type of object this subflow is applicable to, e.g., 'image' or 'video'."
          },
          "input_params": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Input Params",
            "description": "List of input parameters specific to this subflow, replacing the original flow's parameters.",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "display_name",
          "type"
        ],
        "title": "SubFlow",
        "description": "A SubFlow modifies or extends a Flow by overwriting certain parameters like display_name and input_params."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}